requirements.txt
discord.py
sentry_sdk
pymongo[tls]